move mappers and transformations to one file mapper for chat screen
