+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/screen/chat/ChatInteractionListener.kt
@@ -2,7 +2,6 @@
 
 package net.thechance.mena.core_chat.presentation.screen.chat
 
-import net.thechance.mena.core_chat.presentation.screen.chat.model.TextMessageUiState
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
